fix: allow forwarded props to be of any type
Previously, the forwarded props were required to be a truthy object. This makes sense if you are
composing a render prop component and passing `this.state` as the forwarded props. However, you
should be able to pass single primitive values as the forwarded prop - a number, string, boolean,
etc. This commit lifts the type restriction on the second parameter, `newProps`.